pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "C90"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
addopts = "--cov=gendiff --cov-report=xml"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
